+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features [v0.5.7] - Release Date: 2022-05-10
 
 - added file CO
+- added creditMemoConditionCode in condition co menu
 
 ## Resolved [v0.5.6] - Release Date: 2022-05-09